+++ scripts/test_train.sh
@@ -19,5 +19,5 @@ fi
 
 cd train
 python train.py --batch_size=1 --block_size=32 --max_iters=100 --run_name=ci/sdpa/$CURRENT_COMMIT --data_root=/shared/datasets $LOGGING_FLAG
-python train.py --attention_kernel=power --chunk_size=32 --batch_size=1 --block_size=128 --max_iters=100 --run_name=ci/power/$CURRENT_COMMIT --data_root=/shared/datasets $LOGGING_FLAG
-python train.py --attention_kernel=power --batch_size=1 --block_size=32 --max_iters=100 --run_name=ci/power_att/$CURRENT_COMMIT --data_root=/shared/datasets $LOGGING_FLAG
+python train.py --attention_kernel=power --chunk_size=128 --batch_size=2 --block_size=1024 --max_iters=100 --run_name=ci/power_full/$CURRENT_COMMIT --data_root=/shared/datasets $LOGGING_FLAG
+python train.py --attention_kernel=power --batch_size=2 --block_size=1024 --max_iters=100 --run_name=ci/power_att/$CURRENT_COMMIT --data_root=/shared/datasets $LOGGING_FLAG
